@@ -83,17 +83,19 @@
         "lint:check": "make -f etc/Makefile prettiest flags=-s",
         "//only lifecycle scripts below here": "echo",
         "prepare": "husky",
-        "postinstall": "python3 -m venv .venv && cd tools && node pyrun.mjs python -m pip install -U pip && node pyrun.mjs pip install -r requirements.txt"
+        "postinstall": "git config --local include.path ../.gitconfig && python3 -m venv .venv && cd tools && node pyrun.mjs python -m pip install -U pip && node pyrun.mjs pip install -r requirements.txt"
     },
     "dependencies": {
         "axios": "^1.6.8",
         "bigint-isqrt": "^0.3.2",
         "bigint-mod-arith": "^3.3.1",
+        "chroma-js": "^3.1.2",
         "dompurify": "^3.2.1",
         "interactjs": "^1.10.27",
         "p5": "^1.11.0",
         "temml": "^0.10.31",
         "vue": "^3.5.13",
+        "vue-pick-colors": "^1.8.0",
         "vue-router": "^4.5.0"
     },
     "devDependencies": {
@@ -102,6 +104,7 @@
         "@playwright/browser-firefox": "^1.46.1",
         "@playwright/test": "^1.46.1",
         "@tsconfig/node20": "^20.1.4",
+        "@types/chroma-js": "^3.1.0",
         "@types/jsdom": "^21.1.7",
         "@types/node": "^20.14.10",
         "@types/p5": "^1.7.6",
@@ -118,11 +121,11 @@
         "npm-run-all": "^4.1.5",
         "prettier": "^3.3.3",
         "sass": "^1.77.2",
-        "typescript": "^5.5.4",
-        "typescript-eslint": "^8.5.0",
+        "typescript": "^5.7.3",
+        "typescript-eslint": "^8.23.0",
         "vite": "^6.0.0",
         "vitest": "^1.6.0",
         "vue-eslint-parser": "^9.4.3",
-        "vue-tsc": "^2.1.10"
+        "vue-tsc": "^2.2.0"
     }
 }
