@@ -12,8 +12,8 @@ export const baseConfiguration = {
     /* Run tests in files in parallel */
     fullyParallel: false,
     forbidOnly: !!process.env.CI,
-    /* Retry on CI only */
-    retries: process.env.CI ? 2 : 0,
+    /* Retry more on CI */
+    retries: process.env.CI ? 2 : 1,
     /* Opt out of parallel tests on CI. */
     workers: process.env.CI ? 1 : undefined,
     /* Reporter to use. See https://playwright.dev/docs/test-reporters */
