@@ -1,11 +1,12 @@
 import p5 from 'p5'
 import {markRaw} from 'vue'
 
+import {INVALID_COLOR} from './P5Visualizer'
 import {P5GLVisualizer} from './P5GLVisualizer'
 import {VisualizerExportModule} from './VisualizerInterface'
 import type {ViewSize} from './VisualizerInterface'
 
-import {CachingError} from '@/sequences/Cached'
+import {math, MathFormula, CachingError} from '@/shared/math'
 import type {GenericParamDescription, ParamValues} from '@/shared/Paramable'
 import {ParamType} from '@/shared/ParamType'
 import {ValidationStatus} from '@/shared/ValidationStatus'
@@ -22,7 +23,7 @@ import {ValidationStatus} from '@/shared/ValidationStatus'
 This visualizer interprets a sequence as instructions for a drawing machine,
 often known as a "turtle" after a simple drawing robot built in the 1960s.
 For some domain of possible values, each entry in the sequence determines
-a turn angle and step length.  The visualizer displays the resulting polygonal
+an angle and step length.  The visualizer displays the resulting polygonal
 path.
 
 There are two ways to animate the resulting path:
@@ -32,17 +33,47 @@ being drawn some number of steps per frame, as long as more terms of the
 sequence are available.
 
 2.  By setting non-zero values for the fold and stretch parameters in the
-animation section, the user can set the turn angles and/or step lengths to
-gradually increase or decrease over time, resulting in a visual effect
-reminiscent of protein folding. In this mode, the same fixed number of steps
-of the path are re-drawn in every frame.
+animation section, the user can set the angles and/or step lengths to
+gradually change over time, resulting in a visual effect reminiscent of
+protein folding. In this mode, the same fixed number of stepsvof the path
+are re-drawn in every frame.
 
 ## Parameters
 **/
 
+enum RuleMode {
+    List,
+    Formula,
+}
+
+enum AngleMeaning {
+    Turn,
+    Bearing,
+}
+
+enum AngleUnit {
+    Degrees,
+    Radians,
+}
+
+const formulaSymbols = [
+    'n',
+    'a',
+    'A',
+    's',
+    'm',
+    'M',
+    'f',
+    'b',
+    'x',
+    'y',
+] as const
+// set version
+const formulaSymSet = new Set(formulaSymbols)
+
 const paramDesc = {
     /** md
-- Domain: a list of numbers.  These are the values that
+- **Domain**: a list of numbers.  These are the values that
 that the turtle should pay attention to when appearing as
 entries of the sequence.  Values of the sequence
 not occurring in this list will be skipped.
@@ -55,11 +86,14 @@ have a small domain.)
         default: [0n, 1n, 2n],
         type: ParamType.BIGINT_ARRAY,
         displayName: 'Domain',
-        required: true,
+        required: false,
         description:
             'Sequence values to interpret as rules; entries not '
             + 'matching any value here are skipped.',
         hideDescription: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        level: 0,
         validate: function (dom: bigint[], status: ValidationStatus) {
             const seen = new Set()
             for (const element of dom) {
@@ -76,46 +110,71 @@ have a small domain.)
     },
     /** md
 
-The next four parameters give the instructions for the turtle's path (and how
-it changes from frame to frame). Each one can be a single number, in which
-case it is used for every element of the domain. Or it can be a list of
-numbers the same length as the domain, in which case the numbers correspond in
-order: the first number is used when an entry is equal to the first value in
-the domain, the second number is used for entries equal to the second value,
-and so on. For example, if the "Steps" parameter below is "10", then a segment
-10 pixels long will be drawn for every sequence entry in the domain, whereas if
-the domain is "0 1 2" and "Steps" is "20 10 0", then the turtle will move 20
-pixels each time it sees a sequence entry of 0, 10 pixels for each 1, and it
-won't draw anything when the entry is equal to 2 (but it might turn).
+The following set of parameters give the instructions for the turtle's path
+(and how it changes from frame to frame). Each one can be a single number, in
+which case it is used for every element of **Domain**. Or it can be a list
+of numbers the same length as **Domain**, in which case the numbers correspond
+in order: the first number is used when an entry is equal to the first value
+in **Domain**, the second number is used for entries equal to the second value,
+and so on. For example, if the **Step length(s)** parameter below is "10",
+then a segment 10 pixels long will be drawn for every sequence entry in
+**Domain**, whereas if **Domain** is "0 1 2" and **Step length(s)** is
+"20 10 0", then the turtle will move 20 pixels each time it sees a sequence
+entry of 0, 10 pixels for each 1, and it won't draw anything when an entry
+is equal to 2 (but it might turn).
 
-- Turn angle(s): Specifies (in degrees) how much the turtle should turn for
-each sequence entry in the domain. Positive values turn counterclockwise,
-and negative values clockwise.
+- **Angle(s)**: Determines the bearing angle (in degrees) for the turtle before
+each step (corresponding to one sequence entry in **Domain**). Angles are
+measured so that positive angles run counterclockwise, and negative clockwise.
      **/
-    turns: {
+    angles: {
         default: [30, 45, 60],
         type: ParamType.NUMBER_ARRAY,
-        displayName: 'Turn angle(s)',
-        required: true,
+        displayName: 'Angle(s)',
+        required: false,
         description:
             'An angle (in degrees) or a list of angles, in order '
             + 'corresponding to the sequence values listed in Domain.',
         hideDescription: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        level: 0,
     },
     /** md
-- Step length(s): Specifies (in pixels) how far the turtle should move (and
-draw a line segment as it goes) for each sequence entry in the domain. Note
+- **Angle is...**: specifies whether the angle for each step is taken directly
+  as the new bearing of the turtle (counterclockwise from moving horizontally
+  to the right), or as a turn (counterclockwise change in bearing) from its
+  previous bearing. In either case, the turtle takes on its new bearing
+  _before_ taking its next step.
+    **/
+    angleMeaning: {
+        default: AngleMeaning.Turn,
+        type: ParamType.ENUM,
+        from: AngleMeaning,
+        displayName: 'Angle is...',
+        description:
+            "'Bearing' means that the angle directly gives the turtle's "
+            + "new direction; 'Turn' means the turtle turns that amount ccw.",
+        required: true,
+        level: 1,
+    },
+    /** md
+- **Step length(s)**: Specifies (in pixels) how far the turtle should move (and
+draw a line segment as it goes) for each sequence entry in **Domain**. Note
 negative values (for moving backward) are allowed.
      **/
     steps: {
         default: [20],
         type: ParamType.NUMBER_ARRAY,
         displayName: 'Step length(s)',
-        required: true,
+        required: false,
         description:
             'A length (in pixels), or a list of lengths, in order '
             + 'corresponding to the sequence values listed in Domain.',
         hideDescription: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        level: 0,
     },
     /**
 - animationControls: boolean. If true, show folding controls
@@ -125,16 +184,19 @@ negative values (for moving backward) are allowed.
         type: ParamType.BOOLEAN,
         displayName: 'Animation ↴',
         required: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        level: 0,
     },
     /** md
-- Fold rate(s): Specifies (in units of 0.00001 degree) how each turn angle
+- **Folding rate(s)**: Specifies (in units of 0.00001 degree) how each angle
 changes from one frame to the next. For example, if there is just one entry
-of "5" here, then in each frame of the animation, the turn angle for every
-element of the domain will increase by 0.00005 degree. Similarly, if the
-domain is "0 1" and this list is "200 -100" then in each frame the turn angle
-for 0 entries will increase by 0.002 degrees, and the turn angle for 1 entries
+of "5" here, then in each frame of the animation, the angle for every
+element of **Domain** will increase by 0.00005 degree. Similarly, if
+**Domain** is "0 1" and this list is "200 -100" then in each frame the angle
+for 0 entries will increase by 0.002 degrees, and the angle for 1 entries
 will decrease by 0.001 degree. These increments might seem small, but with
-so many turns in a turtle path, a little goes a long way.
+so many steps in a turtle path, a little goes a long way.
      **/
     folds: {
         default: [0],
@@ -145,15 +207,16 @@ so many turns in a turtle path, a little goes a long way.
             'An angle increment (in units of 0.00001 degree), or list of '
             + 'angle increments in order corresponding to the sequence '
             + 'values listed in Domain. If any are nonzero, the path will '
-            + 'animate, adding the increment(s) to the turn angle(s) before '
-            + 'each frame; these additions accumulate from frame to frame. '
+            + 'animate, adding the increment(s) to the angle(s) in the list '
+            + 'above before each frame; these additions accumulate from '
+            + 'frame to frame. '
             + 'Produces a visual effect akin to protein folding.',
         hideDescription: false,
         visibleDependency: 'animationControls',
         visibleValue: true,
     },
     /** md
-- Stretch rate(s): Specifies (in units of 0.01 pixel) how each step length
+- **Stretch rate(s)**: Specifies (in units of 0.01 pixel) how each step length
 changes from one frame to the next.
      **/
     stretches: {
@@ -170,32 +233,73 @@ changes from one frame to the next.
         visibleDependency: 'animationControls',
         visibleValue: true,
     },
-    /**
-- pathLook: boolean. If true, show path style controls
+    /** md
+- **Stroke width(s)**: Gives the width of the segment drawn for each entry,
+  in pixels.
     **/
-    pathLook: {
-        default: true,
-        type: ParamType.BOOLEAN,
-        displayName: 'Path speed/styling ↴',
+    widths: {
+        default: [1],
+        type: ParamType.NUMBER_ARRAY,
+        displayName: 'Stroke width(s)',
         required: false,
+        description:
+            'A width (in pixels), or list of widths corresponding '
+            + 'in order to the sequence values listed in Domain. '
+            + 'The segments of the path are drawn with the corresponding '
+            + 'widths.',
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        validate: function (widths: number[], status: ValidationStatus) {
+            status.mandate(
+                widths.every(n => n > 0),
+                'must be positive'
+            )
+        },
+        level: 0,
+    },
+    /** md
+- **Stroke color(s)**: One or more colors, corresponding to the elements in
+  **Domain** just as all of the above parameters do. You can add a color by
+  clicking the `+` button, and delete a color by selecting it and pressing
+  the `Delete` key.
+    **/
+    strokeColor: {
+        default: ['#c98787'],
+        type: ParamType.COLOR_ARRAY,
+        displayName: 'Stroke color(s)',
+        required: true,
+        description:
+            'A color or list of colors corresponding to the sequence values '
+            + 'listed in Domain. To remove a color, click it and press '
+            + 'the Delete key.',
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.List,
+        level: 0,
     },
     /** md
 
-(Note that as an advanced convenience feature, useful mainly when the domain is
-large, you may specify fewer entries in one of these lists than there are
-elements in the domain, and the last one will be re-used as many times as
-necessary. Using this feature will display a warning, in case you inadvertently
-left out a value.)
+(Note that as an advanced convenience feature, useful mainly when **Domain**
+has many elements, you may specify fewer entries in any of the above lists
+than there are elements in **Domain**, and the last one will be re-used
+as many times as necessary. Using this feature will display a warning, in
+case you inadvertently left out a value.)
 
-The remaining parameters control the speed and style of the turtle path
-display.
-
-- Turtle speed: a number.  If zero (or if any of the fold or stretch rates
-are nonzero), the full path is drawn all at once. Otherwise, the path drawing
-will be animated, and the Turtle speed specifies the number of steps of the
-path to draw per frame. The drawing continues as long as there are additional
-entries of the sequence to display. The Turtle visualizer has a brake on it
-to prevent lag: this speed cannot exceed 1000 steps per frame.
+- **Background color**: The color of the visualizer canvas.
+     **/
+    bgColor: {
+        default: '#6b1a1a',
+        type: ParamType.COLOR,
+        displayName: 'Background color',
+        required: true,
+    },
+    /** md
+- **Turtle speed**: a number.  If zero (or if any of the **Folding rate(s)**
+  or **Stretch rate(s)** are nonzero), the full path is drawn all at once.
+  Otherwise, the path drawing will be animated, and the **Turtle speed**
+  specifiesthe number of steps of the path to draw per frame. The drawing
+  continues as long as there are additional entries of the sequence to display.
+  The Turtle visualizer has a brake on it to prevent lag: this speed cannot
+  exceed 1000 steps per frame.
      **/
     speed: {
         default: 1,
@@ -204,53 +308,241 @@ to prevent lag: this speed cannot exceed 1000 steps per frame.
         required: false,
         description: 'Steps added per frame.',
         hideDescription: false,
-        visibleDependency: 'pathLook',
-        visibleValue: true,
         validate: function (n: number, status: ValidationStatus) {
-            if (n < 0) status.addError('must non-negative')
-            if (n > 1000) status.addError('the speed is capped at 1000')
+            status.forbid(n < 0, 'must non-negative')
+            status.forbid(n > 1000, 'the speed is capped at 1000')
         },
     },
     /** md
-- Stroke width: a number. Gives the width of the segment drawn for each entry,
-in pixels.
+- **Rule mode**: You may select "List" or "Formula". Generally speaking, the
+  parameters above are used to control the Turtle visualization when this
+  Rule mode parameter has its default value of "List". The parameters below
+  are used when the **Rule mode** is "Formula". (Note that **Background color**
+  and **Turtle speed** are used in either case.) Broadly, in Formula mode,
+  you can enter expressions that calculate the different aspects of each
+  segment of the turtle's path (angle, step length, color, etc.) directly
+  from the sequence entries. In each of these formulas, you may use the
+  following variables, the values of which will be filled in for you:
+
+  `n` The index of the entry in the sequence being visualized.
+
+  `a` The value of the entry.
+
+  `s` The serial number of the step starting from one for the first step
+  that the turtle takes.
+
+  `m` The minimum index of the sequence being visualized. Note that the
+  above definitions mean that `n`, `s`, and `m` are related by `n = m + s - 1`.
+
+  `M` The Maximum index of the sequence being visualized. Note this value
+  may be Infinity for sequences that are defined and can be calculated in
+  principle for any index.
+
+  `f` The frame number of this drawing pass. If you use this variable, the
+  visualization will be redrawn from the beginning on every frame,
+  animating the shape of the path.
+
+  `b` The current bearing of the turtle, in degrees counterclockwise from its
+  initial bearing. Note this value represents the bearing _before_ the
+  new one when the angle formula is being computed, and the resulting new
+  bearing when all of the other formulas are being computed. For example,
+  if you compute the color of the stroke using the `b` variable, it will
+  correctly base the color on the bearing of that stroke (not on the previous
+  one).
+
+  `x` The current x-coordinate of the turtle (before the upcoming step).
+
+  `y` The current y-coordinate of the turtle.
+
+  You may also use the symbol `A` as a function symbol in your code, and it
+  will provide access to the value of the sequence being visualized for any
+  index. For example, the formula `A(n+1) - A(n)` or equivalently `A(n+1) - a`
+  would produce the so-called "first differences" of the sequence.
+
+  Note that as you change parameters in List mode, the formulas will be
+  updated in the background to match for you, but not vice versa: If you
+  switch to List mode, the visualization will change to reflect the last
+  List mode values you set. That's because most changes you might make
+  in Formula mode can't possibly be captured in the simpler List mode. Hence,
+  the Turtle visualizer simply doesn't try to keep the List values in
+  sync with Formula changes.
+
      **/
-    strokeWeight: {
-        default: 1,
-        type: ParamType.INTEGER,
-        displayName: 'Stroke width',
+    ruleMode: {
+        default: RuleMode.List,
+        type: ParamType.ENUM,
+        from: RuleMode,
+        displayName: 'Rule mode',
+        required: true,
+        updateAction: function (newRule: string) {
+            const turtle = this instanceof Turtle ? this : null
+            if (turtle === null) return
+            const rule: RuleMode = parseInt(newRule)
+            if (rule === RuleMode.List) {
+                turtle.params.ruleMode.description = ''
+            } else {
+                turtle.params.ruleMode.description =
+                    'The formulas below may use the following symbols: '
+                    + '`a` - sequence entry, `n` - sequence index, '
+                    + '`s` - step number, `m` - minimum index, '
+                    + '`M` - Maximum index, '
+                    + '`A(...)` -- sequence entry at any index, '
+                    + '`b` - current bearing, `x`,`y` - current position, '
+                    + '`f` - frame number (triggers animation).'
+                // If there are any invalid rules, reset them to their
+                // default values:
+                const freshRules = new Set<string>()
+                let name: RuleParam = 'angles'
+                for (name in ruleParams) {
+                    if (turtle.statusOf[name].invalid()) {
+                        turtle.resetParam(name)
+                        freshRules.add(name)
+                    }
+                }
+                if (freshRules.size) turtle.refreshParams(freshRules)
+            }
+        },
+    },
+    /** md
+- **Angle formula**: an expression to compute the angle for a given step of the
+  turtle's path. Just as in list mode, the **Angle is...** parameter
+  above specifies whether the value of this expression is taken directly as
+  the new Bearing of the turtle, or whether the turtle makes a
+  (counterclockwise) Turn by this angle before taking its next step.
+    **/
+    angleFormula: {
+        default: new MathFormula(
+            '0 <= a < 3 ? 30+15a : undefined',
+            formulaSymbols
+        ),
+        type: ParamType.FORMULA,
+        symbols: formulaSymbols,
+        displayName: 'Angle formula',
+        description: 'Computes the angle for the turtle at each step',
         required: false,
-        validate: function (n: number, status: ValidationStatus) {
-            if (n <= 0) status.addError('must be positive')
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+        level: 0,
+    },
+    /** md
+- **Angle measure**: whether the result of the **Angle formula** should be
+  interpreted in degrees or radians. Note that the selection also gives the
+  units in which the value of the `b` (bearing) variable  that can be used
+  in formulas will be determined.
+    **/
+    angleMeasure: {
+        default: AngleUnit.Degrees,
+        type: ParamType.ENUM,
+        from: AngleUnit,
+        displayName: 'Angle measure',
+        required: true,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+    },
+    /** md
+- **Step formula**: an expression to compute the pixel length of each step
+  of the turtle's path.
+    **/
+    stepFormula: {
+        default: new MathFormula('20'),
+        type: ParamType.FORMULA,
+        symbols: formulaSymbols,
+        displayName: 'Step formula',
+        description: 'Computes the pixel length of each turtle step',
+        required: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+        level: 0,
+    },
+    /** md
+- **Width formula**: an expression to compute the pixel width of each step
+  of the turtle's path.
+    **/
+    widthFormula: {
+        default: new MathFormula('1'),
+        type: ParamType.FORMULA,
+        symbols: formulaSymbols,
+        displayName: 'Width formula',
+        description: 'Computes the pixel width of each turtle step',
+        required: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+        level: 0,
+    },
+    /** md
+- **Color formula**: an expression to compute the color of each step of the
+  turtle's path.
+    **/
+    colorFormula: {
+        default: new MathFormula('#c98787'),
+        type: ParamType.FORMULA,
+        symbols: formulaSymbols,
+        displayName: 'Color formula',
+        description: 'Computes the color of each turtle step',
+        required: false,
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+        level: 0,
+    },
+    /** md
+- **Color chooser**: This color picker does not directly control the display.
+  Instead, whenever you select a color with it, the corresponding color
+  string is inserted in the **Color formula** box.
+    **/
+    colorChooser: {
+        default: '#00d0d0',
+        type: ParamType.COLOR,
+        displayName: 'Color chooser:',
+        required: true,
+        description:
+            'Inserts choice into the Color formula, replacing current '
+            + 'selection therein, if any.',
+        visibleDependency: 'ruleMode',
+        visibleValue: RuleMode.Formula,
+        updateAction: function (newColor: string) {
+            const turtle = this instanceof Turtle ? this : null
+            if (turtle === null) return
+            const cfIn = document.querySelector('.param-field #colorFormula')
+            if (!(cfIn instanceof HTMLInputElement)) return
+            const cf = turtle.tentativeValues.colorFormula
+            const start = cfIn.selectionStart ?? cf.length
+            const end = cfIn.selectionEnd ?? start
+            turtle.tentativeValues.colorFormula =
+                cf.substr(0, start) + newColor + cf.substr(end)
         },
-        visibleDependency: 'pathLook',
-        visibleValue: true,
-    },
-    /** md
-- Background color: The color of the visualizer canvas.
-     **/
-    bgColor: {
-        default: '#6b1a1a',
-        type: ParamType.COLOR,
-        displayName: 'Background color',
-        required: true,
-        visibleDependency: 'pathLook',
-        visibleValue: true,
-    },
-    /** md
-- Stroke color: The color used for drawing the path.
-     **/
-    strokeColor: {
-        default: '#c98787',
-        type: ParamType.COLOR,
-        displayName: 'Stroke color',
-        required: true,
-        visibleDependency: 'pathLook',
-        visibleValue: true,
     },
 } satisfies GenericParamDescription
 
-const ruleParams = ['turns', 'steps', 'folds', 'stretches'] as const
+const formulaParamNames = [
+    'angleFormula',
+    'stepFormula',
+    'widthFormula',
+    'colorFormula',
+] as const
+const ruleParams = {
+    angles: 'angleFormula',
+    steps: 'stepFormula',
+    folds: 'angleFormula',
+    stretches: 'stepFormula',
+    widths: 'widthFormula',
+    strokeColor: 'colorFormula',
+} as const
+
+type FormulaParam = (typeof formulaParamNames)[number]
+type RuleParam = keyof typeof ruleParams
+function checkRule(r: string): r is RuleParam {
+    return r in ruleParams
+}
+// the cast below is necessary b/c TypeScript doesn't type Object.fromEntries
+// tightly (#?!@!)
+const formulaRules = Object.fromEntries(
+    formulaParamNames.map(
+        (fmla: FormulaParam): [FormulaParam, RuleParam[]] => [fmla, []]
+    )
+) as Record<FormulaParam, RuleParam[]>
+for (const [rule, fmla] of Object.entries(ruleParams)) {
+    if (checkRule(rule)) formulaRules[fmla].push(rule)
+}
 
 // How many segments to gather into a reusable Geometry object
 // Might need tuning
@@ -261,21 +553,14 @@ class Turtle extends P5GLVisualizer(paramDesc) {
     static description =
         'Use a sequence to steer a virtual turtle that leaves a visible trail'
 
-    // maps from domain to rotations and steps
-    private rotMap = markRaw(new Map<string, number>())
-    private stepMap = markRaw(new Map<string, number>())
-    private foldMap = markRaw(new Map<string, number>())
-    private stretchMap = markRaw(new Map<string, number>())
-    // private copies of rule arrays
-    private turnsInternal: number[] = []
-    private stepsInternal: number[] = []
-    private foldsInternal: number[] = []
-    private stretchesInternal: number[] = []
-
     // variables recording the path
     private vertices = markRaw([new p5.Vector()]) // nodes of path
+    private pathWidths = markRaw([1])
+    private pathLengths = markRaw([0])
+    private pathBearings = markRaw([math.evaluate('0')])
+    private pathColors = markRaw([INVALID_COLOR])
     private chunks: p5.Geometry[] = markRaw([]) // "frozen" chunks of path
-    private bearing = 0 // heading at tip of path
+    private bearing = this.pathBearings[0] // bearing at tip of path
     private cursor = 0 // vertices up to this one have already been drawn
 
     // variables holding the parameter values
@@ -302,8 +587,10 @@ class Turtle extends P5GLVisualizer(paramDesc) {
         const status = super.checkParameters(params)
 
         // lengths of rulesets should match length of domain
-        for (const rule of ruleParams) {
-            const entries = params[rule].length
+        let rule: RuleParam = 'angles'
+        for (rule in ruleParams) {
+            const ruleList: (string | number)[] = params[rule]
+            const entries = ruleList.length
             if (entries > 1 && entries < params.domain.length) {
                 this.statusOf[rule].addWarning(
                     `fewer entries than the ${params.domain.length}-element `
@@ -319,9 +606,14 @@ class Turtle extends P5GLVisualizer(paramDesc) {
             }
         }
 
-        const animating =
+        let animating =
             params.folds.some(value => value !== 0)
             || params.stretches.some(value => value !== 0)
+        if (params.ruleMode === RuleMode.Formula) {
+            animating = formulaParamNames.some(name =>
+                params[name].freevars.has('f')
+            )
+        }
         // warn when animation is turned on for long paths
         // BUG:  when sequence params change this isn't re-run
         // so the warning may be out of date
@@ -369,91 +661,74 @@ class Turtle extends P5GLVisualizer(paramDesc) {
         return status
     }
 
-    storeRules() {
-        // this function creates the internal rule maps from user input
-
-        // create an adjusted internal copy of the rules
-        const ruleParams = [
-            {
-                param: this.turns,
-                local: [0],
-            },
-            {
-                param: this.steps,
-                local: [0],
-            },
-            {
-                param: this.folds,
-                local: [0],
-            },
-            {
-                param: this.stretches,
-                local: [0],
-            },
-        ]
-        ruleParams.forEach(rule => {
-            rule.local = [...rule.param]
-        })
-        // ignore (remove) or add extra rules for excess/missing
-        // terms compared to domain length
-        ruleParams.forEach(rule => {
-            while (rule.local.length < this.domain.length) {
-                rule.local.push(rule.local[rule.local.length - 1])
+    convertTable(fmlaName: FormulaParam) {
+        const primaryRule = formulaRules[fmlaName][0]
+        const frameRule = formulaRules[fmlaName][1]
+        const primaryArray: (string | number)[] = this[primaryRule]
+        const primaryLast = primaryArray.length - 1
+        const dLength = this.domain.length
+        let fmlaSource = ''
+        for (let ix = 0; ix < dLength; ix++) {
+            const key = `"${this.domain[ix]}"`
+            let value = `${primaryArray[Math.min(ix, primaryLast)]}`
+            if (frameRule) {
+                const frameArray = this[frameRule]
+                const frameLast = frameArray.length - 1
+                const frameValue = frameArray[Math.min(ix, frameLast)]
+                if (frameValue) {
+                    const mult =
+                        frameRule === 'stretches' ? '0.01' : '0.00001'
+                    value += ` + ${mult}*f*${frameValue}`
+                }
             }
-            while (rule.local.length > this.domain.length) {
-                rule.local.pop()
+            if (fmlaSource) fmlaSource += ', '
+            else fmlaSource = '{'
+            fmlaSource += `${key}: ${value}`
+        }
+        fmlaSource += '}[string(a)]'
+        return new MathFormula(fmlaSource, formulaSymbols)
+    }
+
+    async parametersChanged(names: Set<string>) {
+        if (this.ruleMode === RuleMode.List) {
+            if (names.has('ruleMode') || names.has('domain')) {
+                // pretend every rule changed
+                for (const rule in ruleParams) names.add(rule)
             }
-        })
-        this.turnsInternal = ruleParams[0].local
-        this.stepsInternal = ruleParams[1].local
-        this.foldsInternal = ruleParams[2].local
-        this.stretchesInternal = ruleParams[3].local
-
-        // create a map from sequence values to rotations
-        for (let i = 0; i < this.domain.length; i++) {
-            this.rotMap.set(
-                this.domain[i].toString(),
-                (Math.PI / 180) * this.turnsInternal[i]
-            )
+            const freshFmlas = new Set<string>()
+            for (const name of names) {
+                if (!checkRule(name)) continue
+                const fmlaName = ruleParams[name]
+                const oldFmla = this[fmlaName].source
+                const newFmla = this.convertTable(fmlaName)
+                if (newFmla.freevars.isSubsetOf(formulaSymSet)) {
+                    this[fmlaName] = newFmla
+                    if (newFmla.source !== oldFmla) {
+                        names.add(fmlaName)
+                        freshFmlas.add(fmlaName)
+                    }
+                    if (
+                        fmlaName === 'angleFormula'
+                        && this.angleMeasure !== AngleUnit.Degrees
+                    ) {
+                        this.angleMeasure = AngleUnit.Degrees
+                        names.add('angleMeasure')
+                        freshFmlas.add('angleMeasure')
+                    }
+                }
+            }
+            if (freshFmlas.size) this.refreshParams(freshFmlas)
         }
-
-        // create a map from sequence values to step lengths
-        for (let i = 0; i < this.domain.length; i++) {
-            this.stepMap.set(this.domain[i].toString(), this.stepsInternal[i])
-        }
-
-        // create a map from sequence values to turn increments
-        // notice if path is static or we are folding
-        this.animating = false
-        for (let i = 0; i < this.domain.length; i++) {
-            // cumulative effect of two ways to turn on folding
-            const thisFold = this.foldsInternal[i]
-            if (thisFold != 0) this.animating = true
-            this.foldMap.set(
-                this.domain[i].toString(),
-                (Math.PI / 180) * (thisFold / this.foldDenom)
-            )
-        }
-
-        // create a map from sequence values to stretch increments
-        // notice if path is static or we are animating
-        // rename folding to animating?
-        for (let i = 0; i < this.domain.length; i++) {
-            if (this.stretchesInternal[i] != 0) this.animating = true
-            this.stretchMap.set(
-                this.domain[i].toString(),
-                this.stretchesInternal[i] / this.stretchDenom
-            )
-        }
+        super.parametersChanged(names)
     }
 
     setup() {
         super.setup()
 
-        // create internal rule maps
-        this.storeRules()
-
         // reset variables
+        this.animating = formulaParamNames.some(fmla =>
+            this[fmla].freevars.has('f')
+        )
         this.firstIndex = this.seq.first
         this.maxLength = this.animating
             ? this.throttleLimit
@@ -471,8 +746,12 @@ class Turtle extends P5GLVisualizer(paramDesc) {
     refresh() {
         // eliminates the path so it will be recomputed, and redraws
         this.vertices = markRaw([new p5.Vector()]) // nodes of path
+        this.pathWidths = markRaw([1])
+        this.pathLengths = markRaw([0])
+        this.pathBearings = markRaw([math.evaluate('0')])
+        this.pathColors = markRaw([INVALID_COLOR])
         this.chunks = markRaw([])
-        this.bearing = 0
+        this.bearing = this.pathBearings[0]
         this.redraw()
     }
 
@@ -480,21 +759,32 @@ class Turtle extends P5GLVisualizer(paramDesc) {
         // blanks the screen and sets up to redraw the path
         this.cursor = 0
         // prepare sketch
-        this.sketch
-            .background(this.bgColor)
-            .noFill()
-            .stroke(this.strokeColor)
-            .strokeWeight(this.strokeWeight)
-            .frameRate(30)
+        this.sketch.background(this.bgColor).noStroke().frameRate(30)
     }
 
-    // Adds the vertices between start and end INCLUSIVE to the current shape
-    addVertices(start: number, end: number) {
-        let lastPos: undefined | p5.Vector = undefined
-        for (let i = start; i <= end; ++i) {
+    // Draws the vertices between start and end INCLUSIVE
+    drawVertices(start: number, end: number) {
+        const sketch = this.sketch
+        let lastPos = this.vertices[start]
+        for (let i = start + 1; i <= end; ++i) {
             const pos = this.vertices[i]
-            if (pos.x !== lastPos?.x || pos.y !== lastPos?.y) {
-                this.sketch.vertex(pos.x, pos.y)
+            const width = this.pathWidths[i]
+            const length = this.pathLengths[i]
+            let radBearing = math.safeNumber(this.pathBearings[i])
+            if (this.angleMeasure === AngleUnit.Degrees) {
+                radBearing *= Math.PI
+                radBearing /= 180
+            }
+            if (pos.x !== lastPos.x || pos.y !== lastPos.y) {
+                sketch.fill(this.pathColors[i])
+                sketch.push()
+                sketch.translate(lastPos.x, lastPos.y).rotateZ(radBearing)
+                if (Math.abs(length) > width / 3) {
+                    sketch.rect(0, -width / 2, length - width / 3, width)
+                    sketch.circle(length - width / 3, 0, width)
+                }
+                sketch.circle(0, 0, width)
+                sketch.pop()
             }
             lastPos = pos
         }
@@ -524,9 +814,7 @@ class Turtle extends P5GLVisualizer(paramDesc) {
         }
         if (drewSome) this.cursor = this.chunks.length * CHUNK_SIZE
         if (this.cursor < newCursor) {
-            sketch.beginShape()
-            this.addVertices(this.cursor, newCursor)
-            sketch.endShape()
+            this.drawVertices(this.cursor, newCursor)
             this.cursor = newCursor
         }
 
@@ -535,12 +823,10 @@ class Turtle extends P5GLVisualizer(paramDesc) {
         if (!this.animating && fullChunks > this.chunks.length) {
             // @ts-expect-error  The @types/p5 package omitted this function
             sketch.beginGeometry()
-            sketch.beginShape()
-            this.addVertices(
+            this.drawVertices(
                 (fullChunks - 1) * CHUNK_SIZE,
                 fullChunks * CHUNK_SIZE
             )
-            sketch.endShape()
             // @ts-expect-error  Ditto :-(
             this.chunks.push(sketch.endGeometry())
         }
@@ -561,25 +847,6 @@ class Turtle extends P5GLVisualizer(paramDesc) {
     // Resulting path should be targetLength steps
     // meaning that vertices.length = that + 1
     extendPath(currentFrames: number, targetLength: number) {
-        // First compute the rotMap and stepMap to use:
-        let rotMap = this.rotMap
-        if (this.animating) {
-            rotMap = new Map<string, number>()
-            for (const [entry, rot] of this.rotMap) {
-                const extra = currentFrames * (this.foldMap.get(entry) ?? 0)
-                rotMap.set(entry, rot + extra)
-            }
-        }
-        let stepMap = this.stepMap
-        if (this.animating) {
-            stepMap = new Map<string, number>()
-            for (const [entry, step] of this.stepMap) {
-                const extra =
-                    currentFrames * (this.stretchMap.get(entry) ?? 0)
-                stepMap.set(entry, step + extra)
-            }
-        }
-
         // Now compute the new vertices in the path:
         this.pathFailure = false
         const len = this.vertices.length - 1
@@ -601,14 +868,78 @@ class Turtle extends P5GLVisualizer(paramDesc) {
                     throw e
                 }
             }
-            const currElementString = currElement.toString()
-            const turnAngle = rotMap.get(currElementString)
-            if (turnAngle !== undefined) {
-                const stepLength = stepMap.get(currElementString) ?? 0
-                this.bearing += turnAngle
-                position.x += Math.cos(this.bearing) * stepLength
-                position.y += Math.sin(this.bearing) * stepLength
+            const input = {
+                n: Number(i),
+                a: Number(currElement),
+                s: Number(i - this.seq.first + 1n),
+                m: Number(this.seq.first),
+                M: Number(this.seq.last),
+                f: currentFrames,
+                b: this.bearing,
+                x: position.x,
+                y: position.y,
+                A: (n: number | bigint) =>
+                    Number(this.seq.getElement(BigInt(n))),
             }
+            let stepLength = 0
+            let clr: unknown = null
+            try {
+                const angle = this.angleFormula.computeWithStatus(
+                    this.statusOf.angleFormula,
+                    input
+                )
+                if (this.statusOf.angleFormula.invalid()) return
+                if (angle !== undefined) {
+                    if (this.angleMeaning === AngleMeaning.Bearing)
+                        this.bearing = angle
+                    else this.bearing = math.add(this.bearing, angle)
+                    input.b = this.bearing
+                    stepLength = math.safeNumber(
+                        this.stepFormula.computeWithStatus(
+                            this.statusOf.stepFormula,
+                            input
+                        ) ?? 0
+                    )
+                    if (this.statusOf.stepFormula.invalid()) return
+                }
+                this.pathWidths.push(
+                    math.safeNumber(
+                        this.widthFormula.computeWithStatus(
+                            this.statusOf.widthFormula,
+                            input
+                        ) ?? 1
+                    )
+                )
+                if (this.statusOf.widthFormula.invalid()) return
+                clr =
+                    this.colorFormula.computeWithStatus(
+                        this.statusOf.colorFormula,
+                        input
+                    ) ?? 0
+                if (this.statusOf.colorFormula.invalid()) return
+            } catch (e) {
+                this.pathFailure = true
+                if (e instanceof CachingError) {
+                    return // need to wait for more elements
+                } else {
+                    // don't know what to do with this
+                    throw e
+                }
+            }
+            let radBearing = math.safeNumber(this.bearing)
+            if (this.angleMeasure === AngleUnit.Degrees) {
+                radBearing *= Math.PI
+                radBearing /= 180
+            }
+            position.x += Math.cos(radBearing) * stepLength
+            position.y += Math.sin(radBearing) * stepLength
+            if (typeof clr === 'string') {
+                this.pathColors.push(this.sketch.color(clr))
+            } else if (math.isChroma(clr)) {
+                this.pathColors.push(this.sketch.color(clr.hex()))
+            } else this.pathColors.push(this.sketch.color('white'))
+            this.pathLengths.push(stepLength)
+            this.pathBearings.push(this.bearing)
             this.vertices.push(position.copy())
         }
     }
